Create config file, `model.json` for our main model
Each ablation test config will stem from this config.
For example, if we want to run a experiment without the dataset cleaning,
you will create a config file named `model_without_cleaning` (based on model.json), change the relevant lines in this config (in this case, the training dataset line), and use this config to run your experiment.